requirements.txt
# ==============================================================================
# NDE 集成学习训练系统 - 依赖列表
# ==============================================================================
# 安装: pip install -r requirements.txt

# ------------------------------------------------------------------------------
# 核心深度学习框架
# ------------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0

# ------------------------------------------------------------------------------
# 科学计算
# ------------------------------------------------------------------------------
numpy>=1.24.0

# ------------------------------------------------------------------------------
# 配置与工具
# ------------------------------------------------------------------------------
pyyaml>=6.0                 # YAML 配置解析
tenacity>=8.0.0             # 重试机制 (数据下载)
pillow>=9.0.0               # 图像处理 (PIL)

# ------------------------------------------------------------------------------
# 可视化与进度
# ------------------------------------------------------------------------------
matplotlib>=3.7.0           # 图表绘制
tqdm>=4.65.0                # 进度条

# ------------------------------------------------------------------------------
# 可选: OOD/Domain 数据生成 (Stable Diffusion)
# 安装: pip install diffusers transformers accelerate
# ------------------------------------------------------------------------------
# diffusers>=0.21.0          # Stable Diffusion 生成
# transformers>=4.30.0       # Hugging Face 模型
# accelerate>=0.20.0         # 分布式推理加速

# ------------------------------------------------------------------------------
# 可选: 高级图像处理 (corruption 数据生成)
# ------------------------------------------------------------------------------
# scipy>=1.10.0             # 部分 corruption 可能需要
# scikit-image>=0.20.0      # 高级图像处理
